requirements.txt
flask>=3.0
flask-cors>=4.0
azure-storage-blob>=12.19
gunicorn>=22.0
python-dotenv>=1.0
bcrypt>=4.1

# Testing
pytest>=8.0
pytest-asyncio>=0.23
playwright>=1.40
requests>=2.31
